feat: unified versioning system and fix all lint/compile errors
Introduce single-source-of-truth versioning with VERSION file, automated
release workflow, and CI-enforced version consistency. Reset all components
to unified v1.0.0. Fix all clippy/ruff/tsc errors.

Versioning:
- Add VERSION file as canonical source (1.0.0)
- Add bump-version.sh and check-version-consistency.sh scripts
- Add release.yml GitHub Actions workflow (tag-triggered)
- Add version-consistency CI job
- Sync operator (7.1.0), scout (0.1.0), web (1.0.0) to 1.0.0
- Standardize doc headers to canonical **Version:** X.Y.Z format
- Add docs/core/versioning.md policy documentation
- Add CHANGELOG.md seeded with v1.0.0 entry
- Web UI reads version dynamically from package.json
- Add make release/version-check/changelog targets

Lint/compile fixes:
- Rust: fix 26 clippy warnings (unused imports/vars, unnecessary casts,
  redundant closures, clamp patterns, dead code, too_many_arguments)
- Python: fix 15 ruff errors (undefined sqlite3, unbound os, unused imports)
- Web: fix TypeScript jest-dom matcher type augmentation
- Fix .gitignore to track AGENTS.md/CLAUDE.md/CHANGELOG.md/docs

+++ operator/src/handlers/api.rs
@@ -2067,9 +2067,9 @@ pub async fn get_trade_latency(
         _ => 24,
     };
 
-    let stats = state.db.get_trade_latency_stats(hours as i32).await?;
+    let stats = state.db.get_trade_latency_stats(hours).await?;
     let histogram =
-        state.db.get_trade_latency_histogram(hours as i32, &[10.0, 50.0, 100.0, 500.0, 1000.0, 5000.0])
+        state.db.get_trade_latency_histogram(hours, &[10.0, 50.0, 100.0, 500.0, 1000.0, 5000.0])
             .await?;
 
     Ok(Json(TradeLatencyResponse {
